+++ src/CleanRegex/Match/AbstractMatchPattern.php
@@ -131,12 +131,12 @@ abstract class AbstractMatchPattern implements MatchPatternInterface, PatternLim
 
     public function flatMap(callable $mapper): array
     {
-        return (new FlatMapper($this->getDetailObjects(), new ArrayMergeStrategy(), $mapper, 'flatMap'))->get();
+        return (new FlatMapper(new ArrayMergeStrategy(), $mapper, 'flatMap'))->get($this->getDetailObjects());
     }
 
     public function flatMapAssoc(callable $mapper): array
     {
-        return (new FlatMapper($this->getDetailObjects(), new AssignStrategy(), $mapper, 'flatMapAssoc'))->get();
+        return (new FlatMapper(new AssignStrategy(), $mapper, 'flatMapAssoc'))->get($this->getDetailObjects());
     }
 
     public function distinct(): array